requirements.txt
streamlit
faiss-cpu
sentence-transformers
requests
unidecode 
rapidfuzz
groq
deep_translator